Adds asc2 "sprites" for all pokemon until gen7 and a system to print'em
Adds asc2 sprites for all pokemon (and an empty sprite), adds a system that receives a "pokebox" and prints it in the terminal. The box has reduced size so it can be visible in the terminal. For now, adds random pokemon and 1 empty slot for testing.
Probably has some memory leaks and lacks comments. 
Also changes de Code Blocks project to a much cleaner makefile :)